backend/requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.32.1
pymongo==4.10.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.18
pydantic==2.10.2
python-dotenv==1.0.1
httpx==0.28.1
bcrypt==4.2.1